+++ configs/Augmentation/__init__.py
@@ -1,1 +1,2 @@
 from .unet_cts_autoaugment import cfg as UNet_CTS_AutoAugment
+from .unet_cts_elastic import cfg as UNet_CTS_Elastic
