+++ configs/__init__.py
@@ -2,6 +2,7 @@ from .CaRTS import CaRTS_AMBF, CaRTS_CTS
 from .TC_CaRTS import TCCaRTS_CTS
 from .Networks import UNet_CTS, HRNet_CTS
 from .Augmentation import UNet_CTS_AutoAugment
+from .Augmentation import UNet_CTS_Elastic
 
 config_dict = {
         "CaRTS_AMBF": CaRTS_AMBF,
@@ -9,5 +10,6 @@ config_dict = {
         "TCCaRTS_CTS": TCCaRTS_CTS,
         "UNet_CTS": UNet_CTS,
         "HRNet_CTS": HRNet_CTS,
-        "UNet_CTS_AutoAugment": UNet_CTS_AutoAugment
+        "UNet_CTS_AutoAugment": UNet_CTS_AutoAugment,
+        "UNet_CTS_Elastic": UNet_CTS_Elastic,
         }
