+++ datasets/augmentation/__init__.py
@@ -1,4 +1,7 @@
 from .autoaugment import AutoAugment
+from .elastic import Elastic
+
 augmentation_dict = {
-    "AutoAugment":AutoAugment,
+    "AutoAugment": AutoAugment,
+    "Elastic": Elastic,
 }
