+++ datasets/causal_tool_seg.py
@@ -74,6 +74,7 @@ class CausalToolSeg(data.Dataset):
             if self.gt_transforms is None:
                 gt = T.ToTensor()(gt)
             else:
+                self.gt_transforms = T.Compose()
                 gt = self.gt_transforms(gt)
             if self.kinematics_transforms is None:
                 kinematics = torch.tensor(kinematics)
